Refactor: Service를 Impl 구조로 변경 (#21)
